pyproject.toml
[project]
name = "langgraph-tutorial-cx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.13",
    "langchain-classic>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-mcp-adapters>=0.2.2",
    "langgraph>=1.1.3",
    "python-dotenv>=1.2.2",
]
